# Replace "weakLinks" with general binary clauses (BinaryImplicationLayeredGraph)