"Layered" here represents subboards, each subboard adds a layer so that links added to parent boards are automatically inherited by subboards.

Deduplication is not done yet, so it's possible a link ends up in both subboard and mainboard if it was added to the subboard first, but we can deal with that along with implementing the other preprocessing passes.

Timings for "PUZZLE NO. 395 : X SUMS - LITTLE KILLER HYBRID SUDOKU!" by akash doulani increased from 131ms to 409ms

Timings for "Oct. 11, 20023: Oct. 10, 2023" by Philip Newman decreased from 78ms to 39ms